REFACTOR : rename db > repository
