+++ src/main/java/bejavawebserver/service/SignUpService.java
@@ -1,6 +1,6 @@
 package bejavawebserver.service;
 
-import bejavawebserver.db.memoryDB;
+import bejavawebserver.repository.memoryDB;
 import bejavawebserver.model.User;
 import org.springframework.stereotype.Service;
